# [CMake] Speed-up Configuration (again)