--- a/applications/plugins/CMakeLists.txt
+++ b/applications/plugins/CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.12)
 
-find_package(Sofa.Framework)
+find_package(Sofa.Config REQUIRED)
 
 if(SOFA_BUILD_TESTS OR SOFA_BUILD_RELEASE_PACKAGE)
     # (Deprecated) Library used to write high level tests involving many components.
@@ -61,7 +61,7 @@ if((${CMAKE_CXX_COMPILER_ID} STREQUAL "GNU") AND (${CMAKE_SYSTEM_NAME} MATCHES "
     sofa_add_subdirectory(plugin SofaPardisoSolver SofaPardisoSolver) # SofaPardisoSolver is only available under linux with gcc
 endif()
 
-find_package(Sofa.GL QUIET)
+sofa_find_package(Sofa.GL QUIET)
 if(Sofa.GL_FOUND)
     sofa_add_subdirectory(plugin SofaCUDA SofaCUDA)           # SofaCUDA plugin can't work without OPENGL
     sofa_add_subdirectory(plugin SofaSimpleGUI SofaSimpleGUI) # SofaSimpleGUI plugin can't work without OPENGL