@alxbilger 2022-10-13
My hero! It was so frustrating to use cmake on SOFA (on Windows). Thanks a lot!

Could you tell me how you profile the cmake configuration?