@fredroy 2022-10-13
> My hero! It was so frustrating to use cmake on SOFA (on Windows). Thanks a lot!
> 
> Could you tell me how you profile the cmake configuration?

https://github.com/sofa-framework/sofa/issues/1838 🥳