@alxbilger 2022-10-13
I confirm the configuration is way faster. Still too slow on Windows though. Thanks